requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
